+++ lib/application/services/url_launcher/url_launcher.dart
@@ -6,8 +6,8 @@ import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:url_launcher/url_launcher.dart';
 
-import '../../../utils/logger.dart';
-import '../../domain/entities/url_launcher.dart';
+import '../../../../utils/logger.dart';
+import '../../../domain/entities/url_launcher.dart';
 
 /// URL起動サービスプロバイダー
 final urlLauncherServiceProvider = Provider.autoDispose(
